Move ingest package tests to ingest package
